gc: rename checkpoint_count to min_checkpoint_count
Because it's the minimal number of checkpoints that must not be deleted,
not the actual number of preserved checkpoints. Do it now, in a separate
patch so as to ease review of the next patch.

While we are at it, fix the comment to gc_set_(min_)checkpoint_count()
which got outdated by commit 5512053fa281 ("box: gc: do not remove files
being backed up").