test: make replication/sync test more robust
The test is very intricate and relies on timings too much so that should
we change anything in the SUBSCRIBE procedure, we could easily get a
spurious test failure. Let's rewrite it a little bit to make it more
robust. The details are in the comment in the code.

While we are at it, let's also fix comments and use test_run.wait_cond
helper for waiting for a condition to come true instead of repeat-sleep
loops.