sql: set names for constant fields within VIEW
If VIEW contains constant fields (e.g. CREATE VIEW v AS SELECT 1, 'k';)
it uses string representation of literal as a field name. In the example
above it would be '1' and 'k'. However, if VIEW is created using AS VALUES
syntax, then expressions representing constant literals lack of names
(since span-expression is not assigned in this case).
Lets generate names for all fields which lack names for VIEW.

Closes #3849