Add package.setpath() to set load paths in app entry point
This patch solves a pressing problem that we have to include a
complicated and non-obvious prologue to many enterprise (and
non-enterprise) apps.

With this patch, users may call package.setpath() in their entry point
script, and the containing directory of that script will be added to
lua load paths, as well as the .rocks subdir.

Without this patch, each app has to handle it on its own.