box: deconstify tuple_format and tuple_field everywhere
tuple_format is a complex reference-counted structure so it isn't
always possible to mark it const in a function argument list even if
the function in fact doesn't modify the format. So far we've managed
to retain const qualifier in a bunch of functions that don't involve
reference counting, however with the introduction of indexed JSON
fields we can't do it anymore, because the underlying json_tree
structure doesn't use const qualifier at all. So let's drop const
qualifier everywhere to avoid annoying type conversions.

Needed for #1012