vinyl: rename dump/compact in/out to input/output
'in' is a reserved keyword in Lua, so using 'in' as a map key was a bad
decision - one has to access it with [] rather than simply with dot.
Let's rename 'in' to 'input' and 'out' to 'output' both in the output
and in the code.

(cherry picked from commit 8f7f7c0317cef74c3c56e4c814383b71485f84c8)