vinyl: ignore unknown .run, .index and .vylog keys
Currently, if we encounter an unknown key while parsing a .run, .index,
or .vylog file we raise an error. As a result, if we add a new key to
either of those entities, we will break forward compatibility although
there's actually no reason for that. To avoid that, let's silently
ignore unknown keys, as we do in case of xrow header keys.

(cherry picked from commit 7bd128ae62de561ce9f851985eeee03b8e10b1cb)