tuple: fix on-stack buffer allocation in tuple_hash_field
The buffer is defined in a nested {} block. This gives the compiler the
liberty to overwrite it once the block has been executed, which would be
incorrect since the content of the buffer is used outside the {} block.
This results in box/hash and viny/bloom test failures when tarantool is
compiled in the release mode. Fix this by moving the buffer definition
to the beginning of the function.

Fixes commit 0dfd99c4a2f4 ("tuple: fix hashing of integer numbers").