sql: rework sqlAllocSrcList to set diag
Refactored sqlAllocSrcList routine as sql_src_list_new and
reworked it to use diag_set in case of memory allocation error.
This will ensure that the sqlSrcListAppend function throws an
error using diag in subsequent patches.

This patch refers to a series of preparatory patches that provide
the use of Tarantool errors in the call tree that includes
sqlNormalizeName, since this call can later return errors.

This patch is not self-sufficient, its usage in sqlSrcListAppend
remained to be non-Tarantool (for now). It means, that if
sql_src_list_new fails in sqlSrcListAppend and sets a diag, it is
never thrown to a user (now).

Part of #3931

(cherry picked from commit dfd2b0fb50eec94f142b19334673c867e695fcac)