sql: simplify VIEW creation procedure
Common table and view are created in a similar two-step way: firstly
called sqlStartTable() to allocate and initialize struct space_def
(which will hold meta-information about space to be created) and provide
set of checks (e.g. test if space already exists, if name is valid
identifier etc); then called sqlEndTable() which emits byte-code to make
insertion to _space, _index, _fk_constraint and so on. On the other
hand, VIEW can feature neither indexes nor constraints. As a result, it
makes no sense to invoke whole procedure of table creation. In other
words, we can short it up to emitting byte-code for single insertion to
_space system space.