checkpoint_daemon: remove useless mtime check
Before making a checkpoint, we check if the last checkpoint was created
(or touched) at least checkpoint_interval seconds ago. There's no point
in such a check as checkpoint daemon cannot be spuriously woken up, only
on timeout. Moreover, this check is actuall broken, because it uses
string.format(%d) for formatting a checkpoint signature while the latter
can be cdata, in which case the checkpoint daemon will be permanently
broken with:

  checkpoint_daemon.lua:49: bad argument #2 to 'format' (number expected, got cdata)

That said, let's remove this useless check altogether.

Closes #4030